bike_app/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
docutils==0.16
Flask==1.1.2
Flask-WTF==0.14.3
geographiclib==1.50
geopy==2.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
pandas==1.2.3
python-dateutil==2.8.1
python-geohash==0.8.5
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
